refactor: installer types and methods
